deliveroo-clone/django-backend/pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["Kiskadee <Kiskadee.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "django_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-dry-rest-permissions = "^1.2.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
